+++ exporters/otlp/src/otlp_http_client.cc
@@ -83,10 +83,10 @@ public:
                                       const std::string &body) noexcept
   {
     std::stringstream ss;
-    ss << "Status:" << response.GetStatusCode() << "Header:";
+    ss << "Status:" << response.GetStatusCode() << ", Header:";
     response.ForEachHeader([&ss](opentelemetry::nostd::string_view header_name,
                                  opentelemetry::nostd::string_view header_value) {
-      ss << "\t" << header_name.data() << " : " << header_value.data() << ",";
+      ss << "\t" << header_name.data() << ": " << header_value.data() << ",";
       return true;
     });
     ss << "Body:" << body;
@@ -116,8 +116,7 @@ public:
         OTEL_INTERNAL_LOG_ERROR("OTLP HTTP Client] Export failed, " << log_message);
         result = sdk::common::ExportResult::kFailure;
       }
-
-      if (console_debug_)
+      else if (console_debug_)
       {
         if (log_message.empty())
         {