+++ CHANGELOG.md
@@ -15,7 +15,10 @@ Increment the:
 
 ## [Unreleased]
 
+* [Metrics SDK] Add support for Pull Metric Exporter [#1701](https://github.com/open-telemetry/opentelemetry-cpp/pull/1701)
+  which includes breaking change in the Metrics api.
 * [BUILD] Add CMake OTELCPP_MAINTAINER_MODE [#1650](https://github.com/open-telemetry/opentelemetry-cpp/pull/1650)
+* [SEMANTIC CONVENTIONS] Upgrade to version 1.14.0 [#1697](https://github.com/open-telemetry/opentelemetry-cpp/pull/1697)
 
 ## [1.6.1] 2022-09-22